user-service/requirements.txt
fastapi
uvicorn
asyncpg
pydantic
python-jose
passlib
email-validator
httpx
bcrypt==4.0.1